Moved validate_query_NIdata to brainImagingSamples. Fixed typescript error restoring ref key, as it is used
